requirements.txt
# ✅ Airflow Core & providers
apache-airflow==2.7.0
apache-airflow-providers-common-sql==1.7.1
apache-airflow-providers-http==4.5.1
apache-airflow-providers-ftp==3.5.1
apache-airflow-providers-imap==3.3.1
apache-airflow-providers-sqlite==3.4.3

# ✅ ETL & pipelines
alembic==1.12.0
psycopg2-binary==2.9.9
sqlalchemy==1.4.49
dbt-core==1.6.6
dbt-postgres==1.6.6

# ✅ Streaming / Kafka
kafka-python==2.0.2

# ✅ Cassandra driver
cassandra-driver==3.28.0

# ✅ Spark support
pyspark==3.4.1
py4j==0.10.9.7

# ✅ Data validation, request handling, and more
requests==2.31.0
pydantic==1.10.12
PyYAML==6.0.1
tenacity==8.2.3

# ✅ Optional: logging / rich CLI
colorlog==4.8.0
rich==13.5.2

# ✅ Required by Airflow or Flask
Flask==2.2.5
Flask-AppBuilder==4.3.3
